# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.5.0",
    "title": "Kubero",
    "description": "Kubero is a web-based tool deploy applications on a Kubernetes clusters. It provides a simple and intuitive interface to manage your clusters, applications, and pipelines."
  },
  "servers": [
    {
      "url": "http://localhost:2000/api"
    }
  ],
  "paths": {
...
    "/cli/apps": {
      "post": {
        "tags": [
          "CLI"
        ],
        "summary": "Create a new app",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
        ]
      },
      "get": {
        "tags": [
          "CLI"
        ],
        "summary": "Get a list of running apps",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": {
              "type": "http",
              "scheme": "bearer",
              "bearerFormat": "JWT"
            }
          }
        ]
      }
    },
...
    "/logs/{pipeline}/{phase}/{app}": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "Get logs for a specific app",
        "description": "Get logs for a specific app",
        "parameters": [
          {
            "name": "pipeline",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Pipeline name"
          },
          {
            "name": "phase",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Phase name"
          },
          {
            "name": "app",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "App name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/events": {
      "get": {
        "tags": [
          "Logs"
        ],
        "summary": "Get the Kubero Kubernetes events",
        "description": "",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}